Remove instances of QRectF
QRectF was used in handling the selection of edges, control points,
and handles. This should not be handled by the library anyway, but
implemented in the editor. It was also used to cull patches outside
the viewport when preparing them for tessellation, which we now no
longer do. Ideally we would reimplement this, but I suspect it does
not have a significant impact on performance in most cases.