Fix edge finding bug
When finding the edge nearest a given point we naively looped over all
edges in the mesh. However, when T-junctions are involved some edges
are part of a tree of edges. Only the leaves of such a tree are truly
adjacent to a patch and available for further editing such as
splitting or subdividing.

We now loop over the patches in the mesh instead and inspect it's four
edges. The computation of the distance between a point and an edge is
still rather inaccurate. Additionally, when finding the nearest edge to
a point contained in a patch we should prefer edges adjacent to that
patch. We may improve upon this in the future, but the current
implementation is good enough for now.