Switch to recoloring patch corners instead of edges
Parent edges along splits do not contribute to the mesh's
colors. Previously we looped over all edges when recoloring, including
these irrelevant parent edges. The palette was thus calculated with
the colors of these parent edges in mind, which led to incorrect
results.

We now loop over the patches in a mesh instead. Each of the four
corners of a patch corresponds to a color actually present in the
mesh. An alternative approach may be to loop over the edges as before,
only now while skipping the irrelevant parent edges.