Add support for recoloring based on Bezier patches
Patch matrices in the library are generally in Hermite form. These are
easier to deal with in some ways, but only the corners of a Hermite
matrix correspond to control points in RGBXY-space. The other elements
are vectors in RGBXY-space and can therefore not easily be recolored.

However, leaving these vectors unaltered when recoloring can cause
issues. Specifically, the color of a patch's corners may have been
correctly modified but the unchanged vectors lead to undesirable
interpolation within the patch.

This problem can perhaps be solved by converting the patch matrices to
Bezier form. All elements of a Bezier matrix correspond to control
points in RGBXY-space. The recoloring can thus be done on a Bezier
patch in its entirety. When the patch is then converted back to
Hermite form its vectors should also have changed appropriately.
Whether this actually leads to desirable results is yet to be
determined.