Fix handle data retrieval issue
Previously we skipped over handles whose edge had an enabled recolored
flag. The issue with this is that the edge stored in a handle is not
necessarily one of the edges controlled by that handle. This means
that we skipped handles that did not belong to recolored edges and
failed to skip handles that did.

The fix works by looping over the edges instead of the handles
directly. We can then retrieve the pair of handles stored in each the
edge provided it is not recolored. When doing this we do need to be
careful not to push back the same pair of handles multiple times since
twin edges share their handles.