Project palette input points onto RGB cube
This is not necessary when the palette is constructed from patch
corners only since these should already be inside the RGB
cube. However, when using interpolated colors as the input for palette
construction it is possible for some colors to go out of bounds. This
is a guaranteed source of reconstruction error since the palette hull
is always projected onto the RGB cube. These out of bounds colors are
clipped anyway when they are drawn on screen, so it seems reasonable
to do this as well when constructing the palette.