Stop projecting hull every simplification step
We used to project the hull onto the RGB cube immediately after its
initial construction and every time after removing an edge. This
contradicts the official description of the ASAP and RGBXY
methods. Instead, we now project only before returning the final
simplified hull. The RMSE for a given hull is also computed using a
projected copy, since projection is the only potential source of
errors.