Fix IO code assuming sequential IDs
When writing a mesh to a file we referred to other parts of the mesh
using their current ID. When reading that file the ID for a handle,
edge, patch, or control point will be equal to its position in the
file. If the original mesh did not have sequential IDs (caused by
removing and adding elements) then this will break. Therefore, instead
of writing the ID of an element we now use its index in the relevant
array which will correspond to its order in the written file.