Recolor continuous t-junctions based on parent curve
Previously we used the opposite curve. This opposite curve may itself
be part of a t-junction, causing a deadlock. The parent curve is also
more representative of how the t-junction's color was originally
created. This does require the inactive parent curve to be kept up to
date, which is enforced by calling fix_inactive_origin_colors().